remote-dev-agent/requirements.txt
python-telegram-bot==20.7
requests>=2.31.0
psutil>=5.9.0
GPUtil>=1.4.0
rich>=13.0.0
pydantic>=2.0.0
watchdog>=3.0.0
python-dotenv>=1.0.0
